server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = "==3.13.*"

[dependency-groups]
dev = [
	"ruff==0.12.2",
]
prod = [
	"pydantic-settings==2.8.1",
	"alembic==1.15.2",
	"python-jose==3.4.0",
	"passlib[argon2]==1.7.4",
	"email-validator==2.2.0",
	"fastapi==0.115.12",
	"asyncpg==0.30.0",
	"uvicorn==0.34.2",
	"python-multipart==0.0.20",
	"sqlmodel==0.0.24",
]

[tool.uv]
default-groups = ["dev", "prod"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
    "migrations",
    "grpc_core",
]

line-length = 120
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = [
    "F401",
    "F403",
]

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

docstring-code-format = false

docstring-code-line-length = "dynamic"
